fix: backend server setup completely

+++ backend/main_api/cmd/init/dev.sh
@@ -1,6 +1,9 @@
 #!/usr/local/bin/zsh
 
-set -x
+if [ -d "backend/main_api/bin" ]; then
+  echo 'Remove backend/main_api/bin before rails new command.'
+  exit 0
+fi
 
 compose/dev.sh run --rm main_api bash -c \
   "set -x \
@@ -10,7 +13,8 @@ compose/dev.sh run --rm main_api bash -c \
     -BCGMST --api --skip --skip-active-storage --skip-gemfile \
     --skip-system-test --skip-turbolinks --skip-webpack-install \
   && bundle exec rails db:create \
-  && bundle exec rails g rspec:install \
+  && bundle exec rails g rspec:install --skip \
   && bundle exec rubocop --auto-gen-config"
 
-rm -rf backend/main_api/log backend/main_api/lib
+mkdir -p backend/main_api/tmp/sockets
+rm -rf backend/main_api/{public,vendor}
